@@ -1,7 +1,7 @@
 [project]
 name = "ytmusicapi"
 description = "Unofficial API for YouTube Music"
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 authors=[{name = "sigma67", email= "ytmusicapi@gmail.com"}]
 license={file="LICENSE"}
 classifiers = [
@@ -70,6 +70,7 @@ files = [
     "ytmusicapi/"
 ]
 mypy_path = "ytmusicapi"
+strict = true
 
 [dependency-groups]
 dev = [
@@ -77,7 +78,7 @@ dev = [
     'sphinx<7',
     'sphinx-rtd-theme',
     "ruff>=0.9.0",
-    "mypy>=1.14.0",
+    "mypy>=1.15.0",
     "pytest>=7.4.4",
     "pytest-cov>=4.1.0",
     "types-requests>=2.31.0.20240218",
