@@ -1,5 +1,5 @@
+import json
 import platform
-from typing import Optional
 
 from requests.structures import CaseInsensitiveDict
 
@@ -7,12 +7,12 @@ from ytmusicapi.exceptions import YTMusicError, YTMusicUserError
 from ytmusicapi.helpers import *
 
 
-def is_browser(headers: CaseInsensitiveDict) -> bool:
+def is_browser(headers: CaseInsensitiveDict[str]) -> bool:
     browser_structure = {"authorization", "cookie"}
     return all(key in headers for key in browser_structure)
 
 
-def setup_browser(filepath: Optional[str] = None, headers_raw: Optional[str] = None) -> str:
+def setup_browser(filepath: str | None = None, headers_raw: str | None = None) -> str:
     contents = []
     if not headers_raw:
         eof = "Ctrl-D" if platform.system() != "Windows" else "'Enter, Ctrl-Z, Enter'"
