@@ -2,13 +2,14 @@
 
 from collections.abc import Iterator
 from contextlib import contextmanager
-from typing import Optional, Protocol
+from typing import Protocol
 
 from requests import Response
 from requests.structures import CaseInsensitiveDict
 
 from ytmusicapi.auth.types import AuthType
 from ytmusicapi.parsers.i18n import Parser
+from ytmusicapi.type_alias import JsonDict
 
 
 class MixinProtocol(Protocol):
@@ -18,15 +19,15 @@ class MixinProtocol(Protocol):
 
     parser: Parser
 
-    proxies: Optional[dict[str, str]]
+    proxies: dict[str, str] | None
 
     def _check_auth(self) -> None:
         """checks if self has authentication"""
 
-    def _send_request(self, endpoint: str, body: dict, additionalParams: str = "") -> dict:
+    def _send_request(self, endpoint: str, body: JsonDict, additionalParams: str = "") -> JsonDict:
         """for sending post requests to YouTube Music"""
 
-    def _send_get_request(self, url: str, params: Optional[dict] = None) -> Response:
+    def _send_get_request(self, url: str, params: JsonDict | None = None) -> Response:
         """for sending get requests to YouTube Music"""
 
     @contextmanager
