@@ -1,5 +1,5 @@
+import typing
 from pathlib import Path
-from typing import Optional, Union
 
 import requests
 
@@ -14,6 +14,7 @@ from ytmusicapi.parsers.library import (
     pop_songs_random_mix,
 )
 from ytmusicapi.parsers.uploads import parse_uploaded_items
+from ytmusicapi.type_alias import JsonDict, JsonList, ParseFuncType, RequestFuncType
 
 from ..auth.types import AuthType
 from ..enums import ResponseStatus
@@ -24,8 +25,8 @@ from ._utils import LibraryOrderType, prepare_order_params, validate_order_param
 
 class UploadsMixin(MixinProtocol):
     def get_library_upload_songs(
-        self, limit: Optional[int] = 25, order: Optional[LibraryOrderType] = None
-    ) -> list[dict]:
+        self, limit: int | None = 25, order: LibraryOrderType | None = None
+    ) -> JsonList:
         """
         Returns a list of uploaded songs
 
@@ -59,10 +60,12 @@ class UploadsMixin(MixinProtocol):
         if results is None:
             return []
         pop_songs_random_mix(results)
-        songs = parse_uploaded_items(results["contents"])
+        songs: JsonList = parse_uploaded_items(results["contents"])
 
         if "continuations" in results:
-            request_func = lambda additionalParams: self._send_request(endpoint, body, additionalParams)
+            request_func: RequestFuncType = lambda additionalParams: self._send_request(
+                endpoint, body, additionalParams
+            )
             remaining_limit = None if limit is None else (limit - len(songs))
             songs.extend(
                 get_continuations(
@@ -73,8 +76,8 @@ class UploadsMixin(MixinProtocol):
         return songs
 
     def get_library_upload_albums(
-        self, limit: Optional[int] = 25, order: Optional[LibraryOrderType] = None
-    ) -> list[dict]:
+        self, limit: int | None = 25, order: LibraryOrderType | None = None
+    ) -> JsonList:
         """
         Gets the albums of uploaded songs in the user's library.
 
@@ -94,8 +97,8 @@ class UploadsMixin(MixinProtocol):
         )
 
     def get_library_upload_artists(
-        self, limit: Optional[int] = 25, order: Optional[LibraryOrderType] = None
-    ) -> list[dict]:
+        self, limit: int | None = 25, order: LibraryOrderType | None = None
+    ) -> JsonList:
         """
         Gets the artists of uploaded songs in the user's library.
 
@@ -114,7 +117,7 @@ class UploadsMixin(MixinProtocol):
             response, lambda additionalParams: self._send_request(endpoint, body, additionalParams), limit
         )
 
-    def get_library_upload_artist(self, browseId: str, limit: int = 25) -> list[dict]:
+    def get_library_upload_artist(self, browseId: str, limit: int = 25) -> JsonList:
         """
         Returns a list of uploaded tracks for the artist.
 
@@ -152,8 +155,10 @@ class UploadsMixin(MixinProtocol):
         items = parse_uploaded_items(results["contents"])
 
         if "continuations" in results:
-            request_func = lambda additionalParams: self._send_request(endpoint, body, additionalParams)
-            parse_func = lambda contents: parse_uploaded_items(contents)
+            request_func: RequestFuncType = lambda additionalParams: self._send_request(
+                endpoint, body, additionalParams
+            )
+            parse_func: ParseFuncType = lambda contents: parse_uploaded_items(contents)
             remaining_limit = None if limit is None else (limit - len(items))
             items.extend(
                 get_continuations(
@@ -163,7 +168,7 @@ class UploadsMixin(MixinProtocol):
 
         return items
 
-    def get_library_upload_album(self, browseId: str) -> dict:
+    def get_library_upload_album(self, browseId: str) -> JsonDict:
         """
         Get information and tracks of an album associated with uploaded tracks
 
@@ -205,7 +210,7 @@ class UploadsMixin(MixinProtocol):
         album["duration_seconds"] = sum_total_duration(album)
         return album
 
-    def upload_song(self, filepath: str) -> Union[ResponseStatus, requests.Response]:
+    def upload_song(self, filepath: str) -> ResponseStatus | requests.Response:
         """
         Uploads a song to YouTube Music
 
@@ -251,7 +256,7 @@ class UploadsMixin(MixinProtocol):
         else:
             return response
 
-    def delete_upload_entity(self, entityId: str) -> Union[str, dict]:  # pragma: no cover
+    def delete_upload_entity(self, entityId: str) -> str | JsonDict:  # pragma: no cover
         """
         Deletes a previously uploaded song or album
 
@@ -270,4 +275,4 @@ class UploadsMixin(MixinProtocol):
         if "error" not in response:
             return ResponseStatus.SUCCEEDED
         else:
-            return response["error"]
+            return typing.cast(str, response["error"])
