@@ -1,9 +1,21 @@
+import typing
+from collections.abc import Callable
 from functools import wraps
+from gettext import GNUTranslations
+from typing import ParamSpec, TypeVar
 
 from ytmusicapi.navigation import *
+from ytmusicapi.type_alias import JsonDict, JsonList
+
+P = ParamSpec("P")
+R = TypeVar("R")
+
+if typing.TYPE_CHECKING:
+    pass
 
 
-def parse_menu_playlists(data, result):
+def parse_menu_playlists(data: JsonDict, result: JsonDict) -> None:
+    """performs in-place replacement based on :param:data: in :param:result"""
     menu_items = nav(data, MENU_ITEMS, True)
     if menu_items is None:
         return
@@ -24,16 +36,17 @@ def parse_menu_playlists(data, result):
             result[watch_key] = watch_id
 
 
-def get_item_text(item, index, run_index=0, none_if_absent=False):
+def get_item_text(item: JsonDict, index: int, run_index: int = 0, none_if_absent: bool = False) -> str | None:
     column = get_flex_column_item(item, index)
     if not column:
         return None
     if none_if_absent and len(column["text"]["runs"]) < run_index + 1:
         return None
-    return column["text"]["runs"][run_index]["text"]
+
+    return typing.cast(str, column["text"]["runs"][run_index]["text"])
 
 
-def get_flex_column_item(item, index):
+def get_flex_column_item(item: JsonDict, index: int) -> JsonDict | None:
     if (
         len(item["flexColumns"]) <= index
         or "text" not in item["flexColumns"][index]["musicResponsiveListItemFlexColumnRenderer"]
@@ -41,20 +54,20 @@ def get_flex_column_item(item, index):
     ):
         return None
 
-    return item["flexColumns"][index]["musicResponsiveListItemFlexColumnRenderer"]
+    return typing.cast(JsonDict, item["flexColumns"][index]["musicResponsiveListItemFlexColumnRenderer"])
 
 
-def get_fixed_column_item(item, index):
+def get_fixed_column_item(item: JsonDict, index: int) -> JsonDict | None:
     if (
         "text" not in item["fixedColumns"][index]["musicResponsiveListItemFixedColumnRenderer"]
         or "runs" not in item["fixedColumns"][index]["musicResponsiveListItemFixedColumnRenderer"]["text"]
     ):
         return None
 
-    return item["fixedColumns"][index]["musicResponsiveListItemFixedColumnRenderer"]
+    return typing.cast(JsonDict, item["fixedColumns"][index]["musicResponsiveListItemFixedColumnRenderer"])
 
 
-def get_dot_separator_index(runs):
+def get_dot_separator_index(runs: JsonList) -> int:
     try:
         index = runs.index({"text": " • "})
     except ValueError:
@@ -63,10 +76,16 @@ def get_dot_separator_index(runs):
     return index
 
 
-def parse_duration(duration):
+def parse_duration(duration: str | None) -> int | None:
+    """
+    Parse duration to a value in seconds.
+
+    :param duration: Duration string
+    :return: Duration in seconds
+    """
     # duration may be falsy or a single space: ' '
     if not duration or not duration.strip():
-        return duration
+        return None
     duration_split = duration.strip().split(":")
     for d in duration_split:
         if not d.isdigit():  # For e.g: "2,343"
@@ -76,16 +95,21 @@ def parse_duration(duration):
     return seconds
 
 
-def i18n(method):
+class SupportsLang(typing.Protocol):
+    lang: GNUTranslations
+
+
+def i18n(method: Callable[P, R]) -> Callable[P, R]:
     @wraps(method)
-    def _impl(self, *method_args, **method_kwargs):
-        method.__globals__["_"] = self.lang.gettext
-        return method(self, *method_args, **method_kwargs)
+    def _impl(*args: P.args, **kwargs: P.kwargs) -> R:
+        self = args[0]
+        method.__globals__["_"] = self.lang.gettext  # type: ignore
+        return method(*args, **kwargs)
 
     return _impl
 
 
-def parse_id_name(sub_run):
+def parse_id_name(sub_run: JsonDict | None) -> JsonDict:
     return {
         "id": nav(sub_run, NAVIGATION_BROWSE_ID, True),
         "name": nav(sub_run, ["text"], True),
