@@ -1,8 +1,12 @@
+import re
+
+from ytmusicapi.type_alias import JsonDict, JsonList, ParseFuncDictType
+
 from .podcasts import parse_episode, parse_podcast
 from .songs import *
 
 
-def parse_mixed_content(rows):
+def parse_mixed_content(rows: JsonList) -> JsonList:
     items = []
     for row in rows:
         if DESCRIPTION_SHELF[0] in row:
@@ -46,7 +50,7 @@ def parse_mixed_content(rows):
     return items
 
 
-def parse_content_list(results, parse_func, key=MTRIR):
+def parse_content_list(results: JsonList, parse_func: ParseFuncDictType, key: str = MTRIR) -> JsonList:
     contents = []
     for result in results:
         contents.append(parse_func(result[key]))
@@ -54,7 +58,7 @@ def parse_content_list(results, parse_func, key=MTRIR):
     return contents
 
 
-def parse_album(result):
+def parse_album(result: JsonDict) -> JsonDict:
     album = {
         "title": nav(result, TITLE_TEXT),
         "type": nav(result, SUBTITLE),
@@ -71,7 +75,7 @@ def parse_album(result):
     return album
 
 
-def parse_single(result):
+def parse_single(result: JsonDict) -> JsonDict:
     return {
         "title": nav(result, TITLE_TEXT),
         "year": nav(result, SUBTITLE, True),
@@ -80,7 +84,7 @@ def parse_single(result):
     }
 
 
-def parse_song(result):
+def parse_song(result: JsonDict) -> JsonDict:
     song = {
         "title": nav(result, TITLE_TEXT),
         "videoId": nav(result, NAVIGATION_VIDEO_ID),
@@ -91,7 +95,7 @@ def parse_song(result):
     return song
 
 
-def parse_song_flat(data):
+def parse_song_flat(data: JsonDict) -> JsonDict:
     columns = [get_flex_column_item(data, i) for i in range(0, len(data["flexColumns"]))]
     song = {
         "title": nav(columns[0], TEXT_RUN_TEXT),
@@ -111,7 +115,7 @@ def parse_song_flat(data):
     return song
 
 
-def parse_video(result):
+def parse_video(result: JsonDict) -> JsonDict:
     runs = nav(result, SUBTITLE_RUNS)
     artists_len = get_dot_separator_index(runs)
     videoId = nav(result, NAVIGATION_VIDEO_ID, True)
@@ -131,7 +135,7 @@ def parse_video(result):
     }
 
 
-def parse_playlist(data):
+def parse_playlist(data: JsonDict) -> JsonDict:
     playlist = {
         "title": nav(data, TITLE_TEXT),
         "playlistId": nav(data, TITLE + NAVIGATION_BROWSE_ID)[2:],
@@ -147,7 +151,7 @@ def parse_playlist(data):
     return playlist
 
 
-def parse_related_artist(data):
+def parse_related_artist(data: JsonDict) -> JsonDict:
     subscribers = nav(data, SUBTITLE, True)
     if subscribers:
         subscribers = subscribers.split(" ")[0]
@@ -159,7 +163,7 @@ def parse_related_artist(data):
     }
 
 
-def parse_watch_playlist(data):
+def parse_watch_playlist(data: JsonDict) -> JsonDict:
     return {
         "title": nav(data, TITLE_TEXT),
         "playlistId": nav(data, NAVIGATION_WATCH_PLAYLIST_ID),
