@@ -1,15 +1,16 @@
 from ytmusicapi.parsers.browsing import *
+from ytmusicapi.type_alias import JsonDict
 
 TRENDS = {"ARROW_DROP_UP": "up", "ARROW_DROP_DOWN": "down", "ARROW_CHART_NEUTRAL": "neutral"}
 
 
-def parse_chart_song(data):
+def parse_chart_song(data: JsonDict) -> JsonDict:
     parsed = parse_song_flat(data)
     parsed.update(parse_ranking(data))
     return parsed
 
 
-def parse_chart_artist(data):
+def parse_chart_artist(data: JsonDict) -> JsonDict:
     subscribers = get_flex_column_item(data, 1)
     if subscribers:
         subscribers = nav(subscribers, TEXT_RUN_TEXT).split(" ")[0]
@@ -24,7 +25,7 @@ def parse_chart_artist(data):
     return parsed
 
 
-def parse_chart_trending(data):
+def parse_chart_trending(data: JsonDict) -> JsonDict:
     flex_0 = get_flex_column_item(data, 0)
     artists = parse_song_artists(data, 1)
     index = get_dot_separator_index(artists)
@@ -40,7 +41,7 @@ def parse_chart_trending(data):
     }
 
 
-def parse_ranking(data):
+def parse_ranking(data: JsonDict) -> JsonDict:
     return {
         "rank": nav(data, ["customIndexColumn", "musicCustomIndexColumnRenderer", *TEXT_RUN_TEXT]),
         "trend": TRENDS[nav(data, ["customIndexColumn", "musicCustomIndexColumnRenderer", *ICON_TYPE])],
