@@ -1,5 +1,5 @@
+from gettext import GNUTranslations
 from gettext import gettext as _
-from typing import Any
 
 from ytmusicapi.navigation import (
     CAROUSEL,
@@ -20,14 +20,15 @@ from ytmusicapi.parsers.browsing import (
     parse_video,
 )
 from ytmusicapi.parsers.podcasts import parse_episode, parse_podcast
+from ytmusicapi.type_alias import JsonDict, JsonList
 
 
 class Parser:
-    def __init__(self, language):
+    def __init__(self, language: GNUTranslations) -> None:
         self.lang = language
 
     @i18n
-    def get_search_result_types(self):
+    def get_search_result_types(self) -> list[str]:
         return [
             _("album"),
             _("artist"),
@@ -41,7 +42,7 @@ class Parser:
         ]
 
     @i18n
-    def get_api_result_types(self):
+    def get_api_result_types(self) -> list[str]:
         return [
             _("single"),
             _("ep"),
@@ -49,8 +50,7 @@ class Parser:
         ]
 
     @i18n
-    def parse_channel_contents(self, results: list) -> dict:
-        # type: ignore[name-defined]
+    def parse_channel_contents(self, results: JsonList) -> JsonDict:
         categories = [
             ("albums", _("albums"), parse_album, MTRIR),
             ("singles", _("singles & eps"), parse_single, MTRIR),
@@ -61,7 +61,7 @@ class Parser:
             ("episodes", _("episodes"), parse_episode, MMRIR),
             ("podcasts", _("podcasts"), parse_podcast, MTRIR),
         ]
-        artist: dict[str, Any] = {}
+        artist: JsonDict = {}
         for category, category_local, category_parser, category_key in categories:
             data = [
                 r["musicCarouselShelfRenderer"]
